Clean up build script a bit and make cygwin happy

+++ build_make_fs_config.sh
@@ -11,22 +11,40 @@
 #         BUILD
 #            INSTALL/lib/pkgconfig
 
-cd e2fsprogs
+ROOT_DIR="$(pwd)"
+
+cd "$ROOT_DIR"
+
+if [ ! -d e2fsprogs ] && [ ! -d e2fsprogs/lib ]; then
+	echo "e2fsprogs seems missing!"
+	exit 1
+fi
+
+case "$(uname -s)" in
+	CYGWIN*)
+		EXE=".exe";;
+esac
+
+cd "$ROOT_DIR/e2fsprogs"
 autoreconf -fsi
 rm -rf BUILD
 mkdir BUILD
 cd BUILD
-../configure -prefix="$(pwd)/INSTALL"
+../configure -prefix="$ROOT_DIR/e2fsprogs/BUILD/INSTALL"
+if [ $? != 0 ]; then echo "e2fsprogs: configure failed!"; exit 2; fi
 make libs
+if [ $? != 0 ]; then echo "e2fsprogs: make libs failed!"; exit 2; fi
 make install-libs
+if [ $? != 0 ]; then echo "e2fsprogs: make install-libs failed!"; exit 2; fi
 
-cd ..
-cd ..
+cd "$ROOT_DIR"
 autoreconf -fsi
 rm -rf BUILD
 mkdir BUILD
 cd BUILD
-PKG_CONFIG_PATH="$(pwd)/../e2fsprogs/BUILD/INSTALL/lib/pkgconfig" ../configure
-make make_fs_config
+PKG_CONFIG_PATH="$ROOT_DIR/e2fsprogs/BUILD/INSTALL/lib/pkgconfig" ../configure
+if [ $? != 0 ]; then echo "e2tools: configure failed!"; exit 2; fi
+make make_fs_config$EXE
+if [ $? != 0 ]; then echo "make_fs_config: make failed!"; exit 2; fi
 
-file make_fs_config
+file make_fs_config$EXE
